Create Coding Activity.java
Write a Java program that allows a user to input multiple words.  Your program should stop accepting words when the user enters "STOP".  Store the words in an ArrayList.  The word STOP should not be stored in the list.

Next, print the size of the list, followed by the list itself.

Then, remove the first and last words stored in the list, but only if the list has a length greater than two.  Finally, reprint the list.


Sample Run:
Please enter words, enter STOP to stop the loop.
cup
spoon
fork
bowl
plate
knife
STOP

6
[cup, spoon, fork, bowl, plate, knife]
[spoon, fork, bowl, plate]


Note: For this activity, you must use the class name, U7_L1_Activity_One and the method, main.

Hint: use a while loop to take the user input - remember that this type of loop runs until the condition you specify is no longer met so you can base this off the user input.