Create U7_L3_Activity_One.java
Write a public static void method named shiftRight which takes a single parameter of an ArrayList of String objects. The method should shift every element of the parameter ArrayList one position to the right, and move the last element of the list into the first position.

For example, if the parameter list passed to the method initially prints as [I, am, here], this would become [here, I, am] after the method is executed.

Write your shiftRight method in the U7_L3_Activity_One class.