Create U7_L4_Activity_One.java
Implement the method countInitial which accepts as parameters an ArrayList of Strings and a letter, stored in a String. (Precondition: the String letter has only one character. You do not need to check for this.) The method should return the number of Strings in the input ArrayList that start with the given letter. Your implementation should ignore the case of the Strings in the ArrayList.

Hint - the algorithm to implement this method is just a modified version of the linear search algorithm.