Create U7_L4_Activity_Two.java
Write a public static method named searchLast which implements a modified version of the linear search algorithm on an ArrayList of String objects. Instead of returning the index of the first appearance of the target String in the ArrayList your method should return the index of the last appearance of this String. If the target String is not in the ArrayList your method should return -1.

Write your searchLast method in the U7_L4_Activity_Two class, and ensure that the parameters for the method are an ArrayList of String objects followed by a String.