Create U7_L5_Activity_Two.java
Write a method, public static void selectSort(ArrayList<Integer> list), which implements a selection sort on the ArrayList of Integer objects list. For example, if the parameter list would be printed as [3, 7, 2, 9, 1, 7] before a call to selectSort, it would be printed as [1, 2, 3, 7, 7, 9] after the method call.

Write your selectSort method in the U7_L5_Activity_Two class.