Rename Unit 07/Lesson 6/Coding Activity 1/U7_L6_Activity_Two.java to Unit 07/Lesson 6/Coding Activity 2/U7_L6_Activity_Two.java
Write a method, public static int insertSort(ArrayList<Integer> list), which implements a insertion sort on the ArrayList of Integer objects list. In addition your insertSort method should return an int which represents a statement execution count recording how many times two elements from the ArrayList are compared. For example, if the parameter list prints as [3, 7, 2, 9, 1, 7] before a call to insertSort, it should print as [1, 2, 3, 7, 7, 9] after the method call. This call should return the value 10 since 10 values need to be compared to implement an insertion sort on this array.

Hint - using a for loop and a break statement rather than a while loop for the inner loop in your insertion sort algorithm will allow you to calculate your count of the comparisons made far more easily.

Write your InsertSort method in the U7_L6_Activity_Two class.