Create U8_L1_Activity_One.java
Write a public static method diagSum, which takes a 2d array of int values as a parameter, and returns the sum of the elements in the lead diagonal as an int value. The lead diagonal is defined as the diagonal line of values starting in the top left corner and proceeding one step right and down for each value until either the bottom or right edge of the array is reached. For example, in the array represented below, the numbers in red make up the lead diagonal.

2	4	5	8
1	1	6	5
3	4	5	2

Write your diagSum method in the U8_L1_Activity_One class.