Create runner_U8_L1_Activity_One.java
Use the runner class to test your method but do not add a main method to your U8_L1_Activity_One.java file or your code will not be scored correctly.

Hint - think about what the row/column indices are for each element in the lead diagonal of the array above. What is the link between the two indices for each element?