Create U8_L1_Activity_Two.java
Write a public static method named multTable which returns a 2d array which represents a multiplication table. Each element in the array should be equal to the product (multiplication) of the indices of the cell. The method should take 2 parameters which are equal to the number of rows and number of columns of the multiplication table. For example the call multTable(6, 4) should return the array represented by the table below:

0	0	0	0
0	1	2	3
0	2	4	6
0	3	6	9
0	4	8	12
0	5	10	15

Write your multTable method in the U8_L1_Activity_Two class.