Create TemperatureMonth.java
For this assignment, you will complete 4 methods in the TemperatureMonth class. This class has a single member variable temperature which is a 2-D array of double values. Each value represents the average temperature on a single day. Each row of the matrix represents a single week, and the entire matrix represents a month.

Unlike in the Gregorian calendar used in most western countries, in the calendar system used in this class months always consist of a whole number of weeks, however you may not assume that the number of days in a week is 7, or that the number of weeks in a month is 4. In other words the parameter 2-D array may have any number of rows and columns but each row will be the same length.


You will implement the following methods:

getMaxTemp - Returns a double equal to the highest average daily temperature in the month.

getMinTemp - Returns a double equal to the lowest average daily temperature in the month.

getMonthlyAverage - Returns a double equal to the average temperature for the whole month.

getWeeklyAverages - Returns a 1-d array of doubles, where each double is the average temperature for a week (i.e. the first element is the average of the values in the first row of the temperature array, the second element is the average of the second row etc.)


For example if the matrix of temperatures was as follows:

74.4	63.5	62.8	65.6	69.9	70.9	77.3
75.0	79.5	65.4	79.4	63.8	62.1	77.5
71.1	68.1	62.0	71.5	60.1	76.7	74.1
71.0	61.6	73.2	77.2	66.9	62.7	71.4

Then the returns of these methods should be:

getMaxTemp: 79.5
getMinTemp: 60.1
getMonthlyAverage: 69.81071428571428
getWeeklyAverages: {69.19999999999999, 71.81428571428572, 69.08571428571429, 69.14285714285714}