Create SpecialityCoffee.java
For this coding activity you will need to write a subclass of the Coffee class which you can view in the Coffee.java file. This class has the following attributes: size (String), isSkinny (boolean), shots (int) and type (String). You will not need to edit this class for the exercise.

You will write a SpecialityCoffee class which extends the Coffee class, and has the additional attribute flavor (String). You will write 3 constructors for your class: the first will be a default constructor with no parameters which creates a small non-skinny, single-shot latte with vanilla. The second will take three parameters: size, type and flavor and set the relevant attributes to these values, and all other values to default. The final constructor has 5 parameters representing each of the attributes of a SpecialityCoffee. When you have written the constructors, you should add a toString method which prints the SpecialityCoffee as though it was a Coffee object, but adds "with" and the flavor to the end of this String.