Create SpecialityCoffee.java
In this exercise you will return to your SpecialityCoffee class which you created for the last coding activity (a sample solution to this exercise is already provided but you may paste your own code from the previous lesson if you prefer). This time you will override the Coffee method getPrice which returns the price in cents for a given coffee.

The SpecialityCoffee method getPrice should return the price given by the Coffee method for that object, plus an extra charge for the flavored syrup. This extra charge is 70 cents if the coffee size is "large" or "extra large", and 50 cents otherwise.