Create DoubleCone.java
Write a class which extends the class Cone (which represents a cone with a single scoop of ice cream). Your class will be named DoubleCone, and will represent a cone with two scoops of ice cream. You will need to override all public Cone methods in your subclass, plus you will overload both the constructor and the setFlavor methods.

A summary of the methods in Cone and the methods you will write in DoubleCone is in the table below. You should challenge yourself to rewrite as little code as necessary - try to call existing methods using super and this as much as possible.