Create Farm.java
In this exercise you will debug a class hierarchy, which consists of the classes Farm, Farmer, Animal, Pig, Sheep and Cow.

In this class hierarchy, a Farm object has Animals and a Farmer, while Pigs, Sheeps and Cows are all Animals. (Note that for the purpose of this assignment a Farmer is not an Animal).

Look in particular for mistakes in the class declarations and constructors. When these are all fixed, everything should compile correctly, and the main method in the runner_Farm class should produce output as shown in the sample run. You do not need to debug the runner file, so if it does not compile or run correctly then there is at least one error in one of the other files. Do not add main methods to any of the other files, or your code will not be scored correctly.